# docs/swagger.yaml
info:
  contact: {}
paths:
  /index:
    get:
      description: 返回欢迎信息
      responses:
        "200":
          description: Welcome
          schema:
            type: string
      summary: 欢迎页面
      tags:
      - 首页
  /user/createUser:
    get:
      description: 新增用户
      parameters:
      - description: 用户名
        in: query
        name: name
        type: string
      - description: 密码
        in: query
        name: password
        type: string
      - description: 确认密码
        in: query
        name: repassword
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 添加用户
      tags:
      - 用户模块
  /user/deleteUser:
    get:
      description: 删除用户
      parameters:
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 删除用户
      tags:
      - 用户模块
  /user/findUserByNameAndPwd:
    post:
      description: 通过用户名和密码登录
      parameters:
      - description: 用户名
        in: formData
        name: name
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 登录
      tags:
      - 用户模块
  /user/getUserList:
    get:
      description: 返回所有用户
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 查看所有用户
      tags:
      - 用户模块
  /user/updateUser:
    post:
      description: 修改用户
      parameters:
      - description: id
        in: formData
        name: id
        type: string
      - description: name
        in: formData
        name: name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: code","message"}
          schema:
            type: string
      summary: 修改用户
      tags:
      - 用户模块
swagger: "2.0"